+++ resources/views/bootstrap-4/components/forms/inputs/radio.blade.php
@@ -1,0 +1,15 @@
+<input
+    name="{{ $name }}"
+    type="radio"
+    id="{{ $id }}"
+    value="{{ $value }}"
+    @checked($isChecked)
+    {{ $hasErrors ? 'aria-describedby="validation-' . $name . '-feedback"' : '' }}
+    {{ $attributes->class([
+        'custom-control-input',
+        'is-invalid' => $hasErrors,
+    ]) }}
+/>
+<label class="custom-control-label" for="{{ $id }}">
+    {!! $label !!}
+</label>